refactor: extract DiagnosticsLogStore.swift from BugNarratorDiagnostics.swift
Extracts the `DiagnosticsLogStore` actor (log ring buffer, ~157 lines)
from BugNarratorDiagnostics.swift into its own file. The
`BugNarratorDiagnostics` namespace enum remains in
BugNarratorDiagnostics.swift (it holds `store = DiagnosticsLogStore()`
and log-formatting helpers).

Byte-identical move of lines 3-159. Both files keep `import Foundation`
and `import OSLog`.

BugNarratorDiagnostics.swift: 223 → 66 lines; new file 160 lines.
Tests: 667.
